+++ xyz-txn-handler/src/main/java/com/here/xyz/pub/mapper/IPubMsgMapper.java
@@ -1,0 +1,9 @@
+package com.here.xyz.pub.mapper;
+
+import com.here.xyz.models.hub.Subscription;
+import com.here.xyz.pub.models.PubTransactionData;
+
+public interface IPubMsgMapper {
+
+    String mapToPublishableFormat(final Subscription sub, final PubTransactionData txnData);
+}